requirements.txt
click==6.7
Flask==1.0.2
flask-restful==0.3.6
flask-jsonpify==1.5.0
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
Werkzeug==0.15.3
azure-storage-blob==1.5.0
azure-common==1.1.22
pandas==0.24.2
pyjwt==1.7.1